requirements.txt
python-telegram-bot==13.15
Flask
python-dotenv
yt-dlp
psycopg2-binary
redis
gunicorn
zarinpal
python-dotenv==0.19.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
redis==4.3.4
requests==2.28.1
cryptography==38.0.4
cachetools==5.2.0